Move Unauthenticated tests to karma ...
