+++ karma.conf.js
@@ -23,6 +23,7 @@ module.exports = function (config) {
         files: [
             'node_modules/utf8/utf8.js',
             'browser/toBundle.js',
+            'tests/unauthenticated/*Test.js',
             'tests/*.js'
         ],
 
@@ -58,7 +59,7 @@ module.exports = function (config) {
 
         // level of logging
         // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
-        logLevel: config.LOG_DEBUG,
+        logLevel: config.LOG_INFO,
 
 
         // enable / disable watching file and executing tests whenever any file changes
